Some refactoring around strange pwhistory format corruption bug.
git-svn-id: svn://svn.code.sf.net/p/passwordsafe/code/trunk/pwsafe/pwsafe@2038 1f79f812-37fb-46fe-a122-30589dd2bf55

+++ AddDlg.h
@@ -35,7 +35,7 @@ public:
   time_t m_tttCPMTime;  // Password creation or last changed datetime
   int m_XTimeInt;
   BOOL m_SavePWHistory;
-  int m_MaxPWHistory;
+  size_t m_MaxPWHistory;
   int m_ibasedata;
   uuid_array_t m_base_uuid;
   PWPolicy m_pwp;
